+++ CHANGELOG.md
@@ -9,11 +9,12 @@
 - Adds new signature to `Map.setProjection` that allows to use just `"globe"` and `"mercator"` instead of projection specification object
 - Adds new option to `Map.setProjection` that allows it to persist the projection even after style change (like `Map.enableGlobeProjection` and `Map.enableMercatorProjection` do)
 - Adds new method `Map.forgetPersistedProjection` that forgets a persisted projection
-- Fixes a bug where space would not load correctly if a new remote style spec was undefined.
 
 ### 🐛 Bug Fixes
 - Fixes a bug where `map.getProjection()` did not return a value when default projection was used.
 - Fixes bugs in space and halo where precedence of constructor options vs remote styles was not enforced correctly.
+- Fixes a bug where space would not load correctly if a new remote style spec was undefined.
+- Fixes some bugs in `AnimatedRouteLayer` and `animation-helpers.ts` where the first and final points in a list of coordinates were removed during path smoothing.
 - Fixes a bug on Safari where requestIdleCallback is not supported.
 
 ### ⚙️ Others